Add principal to new context input on workflow_dispatch.
After msys2/msys2-autobuild#64 autobuild's build and maint workflows
accept a 'context' input, and log all inputs as part of the 'schedule'
job.  That allows this code to pass along information to be logged along
with the job, at this point just the principal (user) who made the
request.  That just leaves the 'cancel' action only logged in the server
error log.